Test broker with many triggers with any channel
